[#50] Command becomes null...
This fix prevents a possible race-condition between the consumption of
messages from the commandQueue and a BluetoothGattCallback being invoked by
the Android platform.

It simply checks to see if the commandQueue is empty, is so then an
EMPTY_COMMAND is returned, otherwise the commandQueue is polled for the next
command.

Finally, ensure that GATT callback operations are internally synchronized
against OperationImpl.

-=david=-